requirements.txt
pydantic
langgraph
langchain
langchain-core
langchain-openai
typing-extensions
tavily-python
python-frontmatter